+++ src/ui/components/translate-button.js
@@ -160,12 +160,11 @@ async function setLiveTranslatedButton(word, sourceLang, targetLang, translateBu
             console.log(resultOfLiveTranslation);
         }
 
-        translateButton.style.color = secondaryColor;
-
         if (resultOfLiveTranslation.length > maxLengthForResult)
             resultOfLiveTranslation = resultOfLiveTranslation.substring(0, maxLengthForResult - 3) + '...';
 
         translateButton.innerHTML = resultOfLiveTranslation;
+        translateButton.style.color = secondaryColor;
         // translateButton.setAttribute('title', resultOfLiveTranslation);
 
         /// Create origin language label
@@ -175,8 +174,7 @@ async function setLiveTranslatedButton(word, sourceLang, targetLang, translateBu
         if (originLanguage !== null && originLanguage !== undefined && originLanguage !== '') {
             langLabel = document.createElement('span');
             langLabel.textContent = originLanguage;
-            langLabel.setAttribute('style', `opacity: 0.7; position: relative; right: -${originLabelPadding}px; bottom: -2.5px; font-size: ${originLabelWidth}px;`)
-            langLabel.style.color = getTextColorForBackground(configs.tooltipBackground.toLowerCase());
+            langLabel.setAttribute('style', `opacity: 0.7; position: relative; right: -${originLabelPadding}px; bottom: -2.5px; font-size: ${originLabelWidth}px; color: ${getTextColorForBackground(configs.tooltipBackground.toLowerCase())} !important`)
             translateButton.appendChild(langLabel);
         }